Moving DataMasking to a private folder
