+++ aws_lambda_powertools/utilities/_data_masking/provider/__init__.py
@@ -1,0 +1,5 @@
+from aws_lambda_powertools.utilities._data_masking.provider.base import BaseProvider
+
+__all__ = [
+    "BaseProvider",
+]
