+++ aws_lambda_powertools/utilities/_data_masking/provider/base.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any
 
-from aws_lambda_powertools.utilities.data_masking.constants import DATA_MASKING_STRING
+from aws_lambda_powertools.utilities._data_masking.constants import DATA_MASKING_STRING
 
 
 class BaseProvider:
