+++ aws_lambda_powertools/utilities/_data_masking/provider/kms/__init__.py
@@ -1,0 +1,5 @@
+from aws_lambda_powertools.utilities._data_masking.provider.kms.aws_encryption_sdk import AwsEncryptionSdkProvider
+
+__all__ = [
+    "AwsEncryptionSdkProvider",
+]
