--- aws_lambda_powertools/utilities/data_masking/__init__.py
@@ -1,5 +1,0 @@
-from aws_lambda_powertools.utilities.data_masking.base import DataMasking
-
-__all__ = [
-    "DataMasking",
-]
