--- aws_lambda_powertools/utilities/data_masking/provider/__init__.py
@@ -1,5 +1,0 @@
-from aws_lambda_powertools.utilities.data_masking.provider.base import BaseProvider
-
-__all__ = [
-    "BaseProvider",
-]
