--- aws_lambda_powertools/utilities/data_masking/provider/kms/__init__.py
@@ -1,5 +1,0 @@
-from aws_lambda_powertools.utilities.data_masking.provider.kms.aws_encryption_sdk import AwsEncryptionSdkProvider
-
-__all__ = [
-    "AwsEncryptionSdkProvider",
-]
