+++ tests/e2e/data_masking/test_e2e_data_masking.py
@@ -4,8 +4,8 @@ from uuid import uuid4
 import pytest
 from aws_encryption_sdk.exceptions import DecryptKeyError
 
-from aws_lambda_powertools.utilities.data_masking import DataMasking
-from aws_lambda_powertools.utilities.data_masking.provider.kms.aws_encryption_sdk import (
+from aws_lambda_powertools.utilities._data_masking import DataMasking
+from aws_lambda_powertools.utilities._data_masking.provider.kms.aws_encryption_sdk import (
     AwsEncryptionSdkProvider,
     ContextMismatchError,
 )
